esp_lvgl_port: add semaphore to verify task stop
Use a sempaphore with the LVGL task so that task stoppage can be
verified in the deinit method. This change adds a new member named
'task_stop_to_ms' to lvgl_port_cfg_t that controls how long to wait
for the task to stop before returning a timeout error. This value
defaults to 3000.